Add version of vector file with commented part
Add version of vector file which should be the basis for a
vector reconstruction working on two axis (dual tilt). It has
a commented part, and data1 should not be taken into account,
because this would refer to the third axis. If this works,
a new version removing the commented part should be done for
vector2 reconstruction, allowing to perform the reconstruction
with only two tilt series. Each one of the tilt series being:
[log(Id+/Io) - log(Id-/Io)]